Rename apply_unitary.py to apply_unitary_protocol.py
