+++ cirq/protocols/has_unitary.py
@@ -23,7 +23,7 @@ import numpy as np
 from typing_extensions import Protocol
 
 from cirq.protocols import qid_shape_protocol
-from cirq.protocols.apply_unitary import ApplyUnitaryArgs
+from cirq.protocols.apply_unitary_protocol import ApplyUnitaryArgs
 from cirq.protocols.decompose import _try_decompose_into_operations_and_qubits
 from cirq import linalg
