+++ cirq/protocols/unitary.py
@@ -25,7 +25,7 @@ from typing_extensions import Protocol
 
 from cirq import linalg
 from cirq.protocols import qid_shape_protocol
-from cirq.protocols.apply_unitary import ApplyUnitaryArgs, apply_unitaries
+from cirq.protocols.apply_unitary_protocol import ApplyUnitaryArgs, apply_unitaries
 from cirq.protocols.decompose import _try_decompose_into_operations_and_qubits
 from cirq.type_workarounds import NotImplementedType
